+++ typescript/infra/config/environments/mainnet3/supportedChainNames.ts
@@ -64,6 +64,7 @@ export const mainnet3SupportedChainNames = [
   'solaxy',
   'somnia',
   'soneium',
+  'sonic',
   'sonicsvm',
   'soon',
   'stable',