+++ typescript/deploy-sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @hyperlane-xyz/deploy-sdk
 
+## 8.1.2
+
+### Patch Changes
+
+- @hyperlane-xyz/aleo-sdk@41.3.1
+- @hyperlane-xyz/cosmos-sdk@41.3.1
+- @hyperlane-xyz/radix-sdk@41.3.1
+- @hyperlane-xyz/sealevel-sdk@41.3.1
+- @hyperlane-xyz/utils@41.3.1
+- @hyperlane-xyz/starknet-sdk@29.1.6
+- @hyperlane-xyz/provider-sdk@8.1.2
+- @hyperlane-xyz/tron-sdk@24.1.5
+
 ## 8.1.1
 
 ### Patch Changes