+++ typescript/metrics/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/metrics
 
+## 0.2.38
+
+### Patch Changes
+
+- Updated dependencies [c201794]
+  - @hyperlane-xyz/sdk@41.3.1
+  - @hyperlane-xyz/utils@41.3.1
+  - @hyperlane-xyz/core@12.1.0
+
 ## 0.2.37
 
 ### Patch Changes