+++ typescript/rebalancer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @hyperlane-xyz/rebalancer
 
+## 27.3.17
+
+### Patch Changes
+
+- Updated dependencies [c201794]
+  - @hyperlane-xyz/sdk@41.3.1
+  - @hyperlane-xyz/metrics@0.2.38
+  - @hyperlane-xyz/utils@41.3.1
+  - @hyperlane-xyz/core@12.1.0
+  - @hyperlane-xyz/provider-sdk@8.1.2
+
 ## 27.3.16
 
 ### Patch Changes